wip: URQL init now seems reliable
I was seeing an error in the browser console that looked like it was including code from @/lib/server, only the string was truncated, e.g. something like require("__webpack....

I've now converted the GraphqlProvider into essentially
a pointer to the component rather than wrapping the components again. This seems to have fixed the issue